# shared/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: The Day I Made That Promise — API de partidas
  version: 1.0.0
  description: >
    Contrato único entre el videojuego (productor), el backend (Spring Boot)
    y los clientes PWA (Angular) y móvil (Flutter). Cualquier cambio a este
    archivo dispara la verificación de compatibilidad en ci-backend.yml.
  license:
    name: Uso académico — UTCH
    url: https://www.utch.edu.mx
servers:
  - url: http://localhost:8080
    description: Local (docker compose)
tags:
  - name: matches
    description: Partidas registradas por el videojuego
  - name: players
    description: Estadísticas agregadas por jugador
paths:
  /api/v1/matches:
    post:
      tags: [matches]
      operationId: registerMatch
      summary: Registra una partida terminada (lo llama el videojuego)
      security:
        - telemetryKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequest'
      responses:
        '201':
          description: Partida registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Falta o es inválida la llave de telemetría
        '409':
          description: La sesión ya fue registrada (envío duplicado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags: [matches]
      operationId: listMatches
      summary: Lista las últimas partidas (PWA y app móvil)
      security: []
      parameters:
        - name: playerId
          in: query
          required: false
          schema: { type: string, maxLength: 64 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        '200':
          description: Partidas, de la más reciente a la más antigua
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchResponse'
        '400':
          description: Parámetros de consulta inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/players/{playerId}/stats:
    get:
      tags: [players]
      operationId: getPlayerStats
      summary: Estadísticas agregadas de un jugador
      security: []
      parameters:
        - name: playerId
          in: path
          required: true
          schema: { type: string, maxLength: 64 }
      responses:
        '200':
          description: Estadísticas del jugador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerStats'
        '404':
          description: El jugador no tiene partidas
components:
  securitySchemes:
    telemetryKey:
      type: apiKey
      in: header
      name: X-Telemetry-Key
  schemas:
    MatchOutcome:
      type: string
      enum: [COMPLETED, FAILED, ABANDONED]
    MatchRequest:
      type: object
      required: [sessionId, playerId, score, durationSeconds, levelReached, progressPercent, outcome, finishedAt, gameVersion]
      properties:
        sessionId:
          type: string
          format: uuid
          description: Identificador único de la partida; evita registros duplicados
        playerId: { type: string, minLength: 1, maxLength: 64 }
        score: { type: integer, minimum: 0 }
        durationSeconds: { type: integer, minimum: 1 }
        levelReached: { type: integer, minimum: 1 }
        progressPercent: { type: integer, minimum: 0, maximum: 100 }
        outcome: { $ref: '#/components/schemas/MatchOutcome' }
        materialsCollected: { type: integer, minimum: 0, default: 0 }
        itemsCrafted: { type: integer, minimum: 0, default: 0 }
        finishedAt: { type: string, format: date-time }
        gameVersion: { type: string, maxLength: 20, example: '0.3.0' }
    MatchResponse:
      allOf:
        - $ref: '#/components/schemas/MatchRequest'
        - type: object
          required: [id, receivedAt]
          properties:
            id: { type: integer, format: int64 }
            receivedAt: { type: string, format: date-time }
    PlayerStats:
      type: object
      required: [playerId, matchesPlayed, bestScore, totalPlaySeconds, highestLevel]
      properties:
        playerId: { type: string }
        matchesPlayed: { type: integer, format: int64 }
        bestScore: { type: integer }
        totalPlaySeconds: { type: integer, format: int64 }
        highestLevel: { type: integer }
        completionRate:
          type: number
          format: double
          description: Porcentaje de partidas con outcome COMPLETED (0–100)
    ApiError:
      type: object
      required: [status, message]
      properties:
        status: { type: integer }
        message: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }